+++ src/Image.cpp
@@ -842,22 +842,29 @@ template class BaseImage<double>;
 template class BaseImage<float>;
 template class BaseImage<int32_t>;
 template class BaseImage<int16_t>;
+template class BaseImage<uint32_t>;
+template class BaseImage<uint16_t>;
 template class BaseImage<std::complex<double> >;
 template class ImageAlloc<double>;
 template class ImageAlloc<float>;
 template class ImageAlloc<int32_t>;
 template class ImageAlloc<int16_t>;
+template class ImageAlloc<uint32_t>;
+template class ImageAlloc<uint16_t>;
 template class ImageAlloc<std::complex<double> >;
 template class ImageView<double>;
 template class ImageView<float>;
 template class ImageView<int32_t>;
 template class ImageView<int16_t>;
+template class ImageView<uint32_t>;
+template class ImageView<uint16_t>;
 template class ImageView<std::complex<double> >;
 template class ConstImageView<double>;
 template class ConstImageView<float>;
 template class ConstImageView<int32_t>;
 template class ConstImageView<int16_t>;
+template class ConstImageView<uint32_t>;
+template class ConstImageView<uint16_t>;
 template class ConstImageView<std::complex<double> >;
-
 } // namespace galsim